Limit GLThread affinity for better graphic performance
BZ: 102384

Set the cpu affinity of the GLThread for the execution of onDrawFrame
and eglSwapBuffers functions. In case of core task switch there may be a
frequency penalty that results in missing frames and a UI jerk.
The old cpu affinity of the GThread is backuped so that if onDrawFrame
spawns threads or forks then the backuped affinity is set to the new threads.
If onDrawFrame changes the cpu affinity the backuped value is reseted
and no cpu affinity is restored when eglSwapBuffers completes.

Change-Id: I41300cd8d61f160eb1621cb313a1eafb3cb18561
Reviewed-on: http://android.intel.com:8080/108116
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Nie, Jun <jun.nie@intel.com>
Reviewed-by: Buhot, Thomas <thomas.buhot@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yin, Fengwei <fengwei.yin@intel.com>
Reviewed-by: Liu, MinX Z <minx.z.liu@intel.com>
Tested-by: Liu, MinX Z <minx.z.liu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>