[PORT FROM MAIN] chk_memalign: improve chk_memalign to match with dlfree
BZ: 135855

chk_memalign mismatch with dlfree, it use a hdr struct adjacent to user
pointer to manage memory but later free this hdr which is not the
original one allocated from dlmalloc before.

This may lead memory corruption and leaks.

One more element has been added to hdr_t to solve this issue. But as the
address of chunk allocated via debug wrappers of dlmalloc should also
always satisfy the MALLOC_ALIGNMENT. In this situation, padding bytes will
be added just ahead of each struct hdr_t and after the head of the chunk.
The padding bytes will be calculated at runtime.

Orig-Change-Id: I09d12ab665ede6bfde1516d9cb30aeb6292993a2
Signed-off-by: wang, biao <biao.wang@intel.com>
Signed-off-by: Chenyang Du <chenyang.du@intel.com>

Change-Id: If63b2c95af1008dc7c39bf0b5dfee79c9f7cd0d8
Reviewed-on: http://android.intel.com:8080/129876
Reviewed-by: Yin, Fengwei <fengwei.yin@intel.com>
Reviewed-by: Du, Chenyang <chenyang.du@intel.com>
Tested-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>