Add mmap64()
This adds mmap64() to bionic so that it is possible to have
large offset passed to kernel. However, the syscall mechanism
only passes 32-bit number to kernel. So effectively, the
largest offset that can be passed is about 43 bits (since
offset is signed, and the number passed to kernel is number
of pages (page size == 4K => 12 bits)).

Change-Id: Ib54f4e9b54acb6ef8b0324f3b89c9bc810b07281
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/176177
Reviewed-by: Lind, Robert <robert.lind@intel.com>
Reviewed-by: Uyyala, Sridhar <sridhar.uyyala@intel.com>
Tested-by: Uyyala, Sridhar <sridhar.uyyala@intel.com>