Add mmap64()
Issue:
BZ: 161435
BMDA-1233

This adds mmap64() to bionic so that it is possible to have
large offset passed to kernel. However, the syscall mechanism
only passes 32-bit number to kernel. So effectively, the
largest offset that can be passed is about 43 bits (since
offset is signed, and the number passed to kernel is number
of pages (page size == 4K => 12 bits)).

Change-Id: I2307b47f2af81b9cacff2bb6d3c38bc4881e5fb8
Category: device enablement
Domain: AOSP-Bionic
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Signed-off-by: sahashau <shaunak.saha@intel.com>
Reviewed-on: https://android.intel.com/212577
Reviewed-by: Sripathi, Srinivas <srinivas.sripathi@intel.com>
Reviewed-by: Kankroliwala, Huzefa NomanX <huzefa.nomanx.kankroliwala@intel.com>
Reviewed-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>
Tested-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>