Bionic: Enable sysenter/sysexit as syscall for x86
BZ: 147898

http://lwn.net/Articles/18411/:
"It all started with an observation that system calls on a modern Pentium
4 processor are far slower than on older CPUs. It seems that, for
whatever reason, software interrupts generated with the int instruction
are very slow with the P4 processor. Since x86 Linux invokes system
calls with "int $0x80", that slowness makes itself felt - especially
with system calls (like getpid()) that would, otherwise, be very fast.
There is an obvious solution to this problem: use the sysenter
instruction instead. sysenter is quite a bit faster on modern Pentium
processors.
...
"
performance improvement:
"The result of all this: the gettimeofday() system call runs in just over
half the time on a P4 processor. The speedup on Pentium 3's is less - a
factor of 1.2 - but is still worthwhile."

performance on atom is about factor of 1.2.

Change-Id: Id772ebf4c00f75179cbe78dcc1f63974aff26779
Category: device enablement
Domain: AOSP-Bionic
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Van De Ven, Arjan <arjan.van.de.ven@intel.com>
Signed-off-by: Du Chenyang <chenyang.du@intel.com>
Signed-off-by: Yin Fengwei <fengwei.yin@intel.com>