Revert "Add mmap64()"
Issue: BMDA-1233

mmap64() was added to bionic so that it is possible to have
large offset passed to kernel. 

However, the syscall mechanism only passes 32-bit value to kernel 
so effectively max offset that can be passed is 43 bits (since
offset is signed, and the number passed to kernel is number
of pages (page size == 4K => 12 bits)).

Use of mmap() as a wrapper to mmap64() could result in wrong
offset calculation for mmap(). mmap() and mmap64() should 
be separate functions to eliminate the offset calc problem.

Change-Id: I48cc9c77d7fc717992dd514bac29fa5be38552ae
Signed-off-by: sahashau <shaunak.saha@intel.com>
Reviewed-on: https://android.intel.com/212571
Reviewed-by: Sripathi, Srinivas <srinivas.sripathi@intel.com>
Reviewed-by: Kankroliwala, Huzefa NomanX <huzefa.nomanx.kankroliwala@intel.com>
Reviewed-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>
Tested-by: Chittim Munirathnam, MadhusudanX <madhusudanx.chittim.munirathnam@intel.com>